Fixed subtle bug in loadDb caused by nested forms.
The loadDb page had a subtle bug - the open <form> tag for the "Load Database"
button was being ignored because of the form tags introduced in
admin_form.djt.html. It seems that if a browser sees an opening form tag while
already in a form it just ignores the tag.

The load database form stil actually worked because the action of the form
introduced in admin_form.djt.html was empty, so loadDb posted to itself anyway
(as it should).

It looks like the logic of the if statement controlling the presense of the
form in admin_form.djt.html is backwards - it includes a form if the action is
empty, rather than when it's not empty. Regardless I've just removed the whole
thing as it doesn't get used anywhere apart from from loadDb which doesn't need
the form.